Simplify array reshaping
